improvements from findbugs
git-svn-id: file:///home/SUBVERSION/jas/trunk@5310 eafcb37d-f50b-0410-b8f7-d4cb9effa109

+++ src/edu/jas/ps/MultiVarPowerSeries.java
@@ -884,7 +884,7 @@ public class MultiVarPowerSeries<C extends RingElem<C>> implements RingElem<Mult
         GenPolynomial<C> p = null;
         int t = 0;
         while (lazyCoeffs.homCheck.get(t)) {
-            p = lazyCoeffs.coeffCache.get(t);
+            p = lazyCoeffs.coeffCache.get((long)t);
             if (p == null || p.isZERO()) { // ??
                 t++;
                 continue;
